# docs/swagger.yaml
basePath: /
definitions:
  schemas.StudentResponse:
    properties:
      active:
        example: true
        type: boolean
      age:
        example: 30
        type: integer
      cpf:
        example: 98765432100
        type: integer
      createdAt:
        example: "2025-03-31T12:00:00Z"
        type: string
      deletedAt:
        description: zero time
        example: "0001-01-01T00:00:00Z"
        type: string
      email:
        example: maria@email.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Maria Oliveira
        type: string
      updatedAt:
        example: "2025-03-31T12:30:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server Student API
  title: Student API
  version: "1.0"
paths:
  /hello:
    get:
      description: Retorna Hello
      produces:
      - text/plain
      responses:
        "200":
          description: hello
          schema:
            type: string
      summary: Hello world
      tags:
      - test
  /students:
    get:
      consumes:
      - application/json
      description: Retrieve students details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.StudentResponse'
        "404":
          description: Not Found
      summary: Get a list of students
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Create and store a new student in the database
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.StudentResponse'
        "400":
          description: Bad Request
      summary: Create student
      tags:
      - students
  /students/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a student from the database using their ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.StudentResponse'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a student
      tags:
      - students
    get:
      consumes:
      - application/json
      description: Retrieve student details using the student's ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.StudentResponse'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get student by ID
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Update a student's information using their ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.StudentResponse'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update an existing student
      tags:
      - students
schemes:
- http
swagger: "2.0"